# docs/architecture-pack/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: Revizr API
  version: "1.0.0"
  description: |
    Revizr backend API — exam revision platform for UK students (KS3, GCSE, A-Level, 11+).
    All endpoints require Bearer JWT authentication unless noted as public.
    Base URL: https://api.revizr.co.uk/v1
  contact:
    name: Revizr Engineering
    email: engineering@revizr.co.uk
  license:
    name: Proprietary
servers:
  - url: https://api.revizr.co.uk/v1
    description: Production (eu-west-2)
  - url: https://api.staging.revizr.co.uk/v1
    description: Staging (eu-west-2)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token issued by POST /auth/login or POST /auth/refresh.
        Expires in 15 minutes. Pass in Authorization header as "Bearer <token>".
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "email is required"
        details:
          type: object
          example: { "field": "email", "constraint": "format" }
...
    Subject:
      type: object
      required: [id, user_id, subject_name, exam_board, level, is_active]
      properties:
        id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        user_id:
          type: string
          format: uuid
          example: "8f274e3a-1c2d-4b5e-9f6a-7d8c9b0a1e2f"
        subject_name:
          type: string
          example: "Mathematics"
        exam_board:
          type: string
          example: "AQA"
        level:
          type: string
          enum: [11plus, ks3, gcse, alevel]
          example: "gcse"
        is_active:
          type: boolean
          example: true
...
    PracticeSession:
      type: object
      required: [id, user_id, subject_id, status, started_at, question_count]
      properties:
        id:
          type: string
          format: uuid
          example: "d4e5f6a7-b8c9-0d1e-2f3a-4b5c6d7e8f9a"
        user_id:
          type: string
          format: uuid
          example: "8f274e3a-1c2d-4b5e-9f6a-7d8c9b0a1e2f"
        subject_id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        status:
          type: string
          enum: [active, completed, abandoned]
          example: "active"
        started_at:
          type: string
          format: date-time
          example: "2026-05-15T09:30:00Z"
        ended_at:
          type: string
          format: date-time
          nullable: true
          example: null
        question_count:
          type: integer
          example: 10
        attempts:
          type: array
          items:
            $ref: "#/components/schemas/QuestionAttempt"
    QuestionAttempt:
      type: object
      required: [id, question_id, presented_at]
      properties:
        id:
          type: string
          format: uuid
          example: "e5f6a7b8-c9d0-1e2f-3a4b-5c6d7e8f9a0b"
        question_id:
          type: string
          format: uuid
          example: "11223344-5566-7788-99aa-bbccddeeff00"
        presented_at:
          type: string
          format: date-time
          example: "2026-05-15T09:30:05Z"
        self_mark_score:
          type: integer
          nullable: true
          example: 2
        time_spent_seconds:
          type: integer
          nullable: true
          example: 47
        mark_scheme_viewed:
          type: boolean
          example: false
    ProgressSummary:
      type: object
      required: [user_id, subject_id, period, topics]
      properties:
        user_id:
          type: string
          format: uuid
          example: "8f274e3a-1c2d-4b5e-9f6a-7d8c9b0a1e2f"
        subject_id:
          type: string
          format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        period:
          type: string
          example: "2026-W20"
        topics:
          type: array
          items:
            $ref: "#/components/schemas/TopicProgress"
    TopicProgress:
      type: object
      required: [topic_tag, score_avg, questions_attempted, questions_correct]
      properties:
        topic_tag:
          type: string
          example: "algebra"
        score_avg:
          type: number
          format: float
          example: 0.72
        questions_attempted:
          type: integer
          example: 14
        questions_correct:
          type: integer
          example: 10
    ChildSummary:
      type: object
      required: [student_id, display_name, subjects]
      properties:
        student_id:
          type: string
          format: uuid
          example: "8f274e3a-1c2d-4b5e-9f6a-7d8c9b0a1e2f"
        display_name:
          type: string
          example: "Amara"
        subjects:
          type: array
          items:
            $ref: "#/components/schemas/Subject"
...
  responses:
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "UNAUTHORIZED"
            message: "Invalid or expired access token"
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "FORBIDDEN"
            message: "You do not have permission to access this resource"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "NOT_FOUND"
            message: "Resource not found"
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "VALIDATION_ERROR"
            message: "Request body failed validation"
...
paths:
...
  /progress:
    get:
      tags: [progress]
      operationId: getProgressSummary
      summary: Get progress summary for the current user
      parameters:
        - name: subject_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        - name: period
          in: query
          required: false
          description: ISO week period, e.g. 2026-W20. Defaults to current week.
          schema:
            type: string
          example: "2026-W20"
      responses:
        "200":
          description: Progress summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgressSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /progress/topics:
    get:
      tags: [progress]
      operationId: getProgressTopics
      summary: Get topic-level progress breakdown
      parameters:
        - name: subject_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        - name: weeks
          in: query
          required: false
          description: Number of recent weeks to include. Default 8.
          schema:
            type: integer
            minimum: 1
            maximum: 52
            default: 8
          example: 8
      responses:
        "200":
          description: Topic-level breakdown
          content:
            application/json:
              schema:
                type: object
                properties:
                  topics:
                    type: array
                    items:
                      $ref: "#/components/schemas/TopicProgress"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─── PARENT ──────────────────────────────────────────────────────────────────
  /parent/children:
    get:
      tags: [parent]
      operationId: listChildren
      summary: List students linked to the current parent account
      description: Requires role=parent.
      responses:
        "200":
          description: Linked children
          content:
            application/json:
              schema:
                type: object
                properties:
                  children:
                    type: array
                    items:
                      $ref: "#/components/schemas/ChildSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /parent/children/{studentId}/progress:
    get:
      tags: [parent]
      operationId: getChildProgress
      summary: Get a linked student's progress (parent view)
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "8f274e3a-1c2d-4b5e-9f6a-7d8c9b0a1e2f"
        - name: subject_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          example: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
      responses:
        "200":
          description: Child's progress summary
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProgressSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /parent/children/{studentId}/sessions:
    get:
      tags: [parent]
      operationId: getChildSessions
      summary: Get a linked student's practice session history
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "8f274e3a-1c2d-4b5e-9f6a-7d8c9b0a1e2f"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
          example: 20
        - name: before
          in: query
          required: false
          description: Cursor — return sessions before this timestamp
          schema:
            type: string
            format: date-time
          example: "2026-05-15T00:00:00Z"
      responses:
        "200":
          description: Session history
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessions:
                    type: array
                    items:
                      $ref: "#/components/schemas/PracticeSession"
                  next_cursor:
                    type: string
                    format: date-time
                    nullable: true
                    example: "2026-05-08T14:30:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /parent/children/{studentId}/controls:
    post:
      tags: [parent]
      operationId: updateParentalControls
      summary: Update parental controls for a linked student
      description: Controls available in v1 — session time limits, daily question caps.
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "8f274e3a-1c2d-4b5e-9f6a-7d8c9b0a1e2f"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                daily_question_cap:
                  type: integer
                  minimum: 0
                  maximum: 200
                  nullable: true
                  description: Max questions per day; null = no cap
                  example: 50
                session_duration_minutes:
                  type: integer
                  minimum: 5
                  maximum: 120
                  nullable: true
                  description: Max single session duration; null = no limit
                  example: 45
      responses:
        "200":
          description: Controls updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  student_id:
                    type: string
                    format: uuid
                    example: "8f274e3a-1c2d-4b5e-9f6a-7d8c9b0a1e2f"
                  daily_question_cap:
                    type: integer
                    nullable: true
                    example: 50
                  session_duration_minutes:
                    type: integer
                    nullable: true
                    example: 45
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── SUBSCRIPTIONS ───────────────────────────────────────────────────────────
...